Another option
This approach allows ALTERNATE_SERVICE_VARAINTS to be defined per-env.
These values, if they exist, are used to expose access to other process queues.
Then an environment-bound task can look for the specific queue it needs, or
use the default if not available.

This feels a bit cleaner than before, as we no longer have to hard-code 'lms'
into settings files.